Introduce experiments
This patch defines a new project structure in order to integrate ML
experiments.
