+++ tensorcraft/backend/views/__init__.py
@@ -1,0 +1,3 @@
+from .model import ModelView
+from .server import ServerView
+from .experiment import ExperimentView
