+++ tensorcraft/experiment.py
@@ -1,0 +1,2 @@
+class Experiment:
+    pass
