+++ tensorcraft/server.py
@@ -7,18 +7,17 @@ import pid
 import semver
 
 import tensorcraft
-import tensorcraft.model
-import tensorcraft.storage.local
 
 from aiojobs.aiohttp import atomic, setup
 from functools import partial
 from typing import Awaitable
 
 from tensorcraft import arglib
-from tensorcraft import handlers
 from tensorcraft import tlslib
+from tensorcraft.backend import views 
+from tensorcraft.backend import model
+from tensorcraft.backend import modelrt
 from tensorcraft.logging import internal_logger
-from tensorcraft.storage import metadata
 
 
 class Server:
@@ -29,7 +28,7 @@ class Server:
                   host: str = None, port: str = None,
                   preload: bool = False,
                   close_timeout: int = 10,
-                  strategy: str = tensorcraft.model.Strategy.No.value,
+                  strategy: str = model.Strategy.No.value,
                   logger: logging.Logger = internal_logger):
         """Create new instance of the server."""
 
@@ -44,28 +43,22 @@ class Server:
 
         # TODO: use different execution strategies for models and
         # fallback to the server-default execution strategy.
-        loader = tensorcraft.model.Loader(strategy=strategy, logger=logger)
+        loader = model.Loader(strategy=strategy, logger=logger)
 
-        # A metadata storage with models details.
-        meta = metadata.DB.new(path=data_root)
-
-        storage = tensorcraft.storage.local.FileSystem.new(
-            path=data_root, meta=meta, loader=loader)
-
-        models = await tensorcraft.model.Cache.new(
-            storage=storage, preload=preload)
+        storage = modelrt.FsStorage.new(path=data_root, loader=loader)
+        models = await model.Cache.new(storage=storage, preload=preload)
 
         self.app = aiohttp.web.Application(client_max_size=1024**10)
 
         self.app.on_startup.append(cls.app_callback(self.pid.create))
         self.app.on_response_prepare.append(self._prepare_response)
-        self.app.on_shutdown.append(cls.app_callback(meta.close))
+        self.app.on_shutdown.append(cls.app_callback(storage.close))
         self.app.on_shutdown.append(cls.app_callback(self.pid.close))
 
         route = partial(route_to, api_version=tensorcraft.__apiversion__)
 
-        models_view = handlers.ModelView(models)
-        server_view = handlers.ServerView(models)
+        models_view = views.ModelView(models)
+        server_view = views.ServerView(models)
 
         self.app.add_routes([
             aiohttp.web.put(
